Manual rule updates PART 8 - linter fixes 2
This is in order to null the `DefaultAttackSequence` on certain actors that have inherited it from `^Infantry`.
Fixed linter errors "Actor type `pres` trait `WithInfantryBody` field `DefaultAttackSequence` references an undefined sequence `shoot` on image `pres`". The linter is a lot stricter than the engine now, so this worked ingame but the linter threw errors. The linter of the previous release also didn't have a problem with it because it was less strict then.

As a side note a lot more civilian actors should require this fix but they have `shoot` removed from their sequences - INCONSISTENCY!